// src/app/openapi-gen/model/todoListNameDTO.ts
/**
 * API for Todo App
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TodoListNameDTO { 
    count?: number;
    listId?: string;
    fromDate?: string;
    toDate?: string;
    listName?: string;
}